Unit-test cleanup SectionQuery now stores the Section enum instead of just a short
--HG--
branch : NewQuery

+++ VisioAutomation_2010/VisioAutomation/ShapeSheet/Query/CellQuery_Columns.cs
@@ -158,7 +158,7 @@ namespace VisioAutomation.ShapeSheet.Query
            public SectionQuery Add(IVisio.VisSectionIndices section)
            {
                int ordinal = items.Count;
-               var sec = new SectionQuery(this.parent, ordinal, (short) section);
+               var sec = new SectionQuery(this.parent, ordinal, section);
                this.items.Add(sec);
                return sec;
            }
